Make volume of SurfPhase consistent with h = u + P*v
Defines molar volume to be zero. Density remains defined as being
per unit area or length, depending on the dimensionality of the phase.

Also fixes setting surface composition using mass/mole fractions so
that the sum of concentrations always equals the site density.

Partially resolves #1312

+++ include/cantera/thermo/LatticeSolidPhase.h
@@ -311,7 +311,7 @@ public:
         throw NotImplementedError("LatticeSolidPhase::getConcentrations");
     }
 
-    virtual doublereal concentration(int k) const {
+    virtual doublereal concentration(size_t k) const {
         throw NotImplementedError("LatticeSolidPhase::concentration");
     }
